+++ src/com/jec/jec-commons-index.ts
@@ -26,6 +26,7 @@ export {DecoratorParser} from "./commons/files/reflect/utils/DecoratorParser";
 export {DecoratorPropertiesBuilder} from "./commons/files/reflect/utils/DecoratorPropertiesBuilder";
 export {ImportRef} from "./commons/files/reflect/utils/ImportRef";
 export {ImportRefParser} from "./commons/files/reflect/utils/ImportRefParser";
+export {InspectMode} from "./commons/files/reflect/utils/InspectMode";
 export {PathStats} from "./commons/files/reflect/utils/PathStats";
 //--> com/jec/commons/files/reflect
 export {DecoratorProperties} from "./commons/files/reflect/DecoratorProperties";