+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - **build**: adding build script to npm instal process for GlassCat instal optimisation
 - **index**: refactoring the module index
+- **@Bootstrap**: adding the new `@Bootstrap` decorator as specified by JEC
 
 <a name="jec-commons-1.1.2"></a>
 ## **1.1.2** (2017-07-14)